Configure and provision DNSMASQ
The provisioning of DNSMASQ is done by a shell script.
PXE is configured to serve a Memtest86+ image out of the box as a
simple mean to validate the setup.

A possible test strategy is to manually create an empty Virtualbox VM
configured in the following way:
1. Only one network interface connected to the default "intnet"
   internal network
2. Boot from network as the first boot choice

Note that this manually created VM should not have other network
interfaces because it could interfere with the DHCP server provided by
DNSMASQ.

This machine cannot not be configured with Vagrant (as of version
1.9.3) because it would have two network interfaces as Vagrant
needs a NATed eth0 interface to talk with each machine, so it
cannot have only one "internal network" interface.